fix: correctness bugs + M1 boundary hardening (from the code review)
Fixes from an adversarial code review across the kernel core and all five arch ports:

- riscv (HIGH): close the syscall-return trap window -- .Lrestore restored mepc before mstatus
  while reached with MIE=1, so an IRQ in that 2-instruction window clobbered mepc and the mret
  jumped back into the epilogue at the caller's privilege. A crash on any syscall racing an IRQ
  on the C6/qemu-virt. Restore mstatus (MIE=0) first.
- esp32c6: time ran 4.17% long -- ns-per-tick truncated 6.25 to 6. Convert with the exact 25/4
  ratio, named (mtime_ticks_to_ns / mtime_ns_to_ticks).
- xtensa: the doorbell dispatcher masked the injected line unconditionally, stranding a tier-2
  (irq_attach) line masked after one delivery. Leave masking to the bound handler (matches sim/
  riscv/rx).
- telemetry: the SESSION arch-id map disagreed between the build injection and record.h/kicktrace
  (RX and Xtensa swapped, RISC-V absent); align to the enum + add riscv=5. Fix telemetry-record.sh
  XMC trace clock 120->144 MHz (SystemCoreClock).
- thread: copy the (possibly user-supplied) name into a bounded TCB buffer; the fault reporter no
  longer %s-derefs a user pointer.
- M1 boundary hardening (arch-neutral, not M2): SlotPool::resolve compares the full generation
  (a handle with a dirty top byte no longer aliases); clock_now rejects a null/misaligned
  out-pointer; thread_spawn rejects a caller stack whose base+size wraps.

+++ kernel/include/kickos/slotpool.h
@@ -71,8 +71,10 @@ namespace kickos
                 return nullptr;
             }
             int const index = handle & ((1 << kIndexBits) - 1);
-            uint16_t const gen = static_cast<uint16_t>(static_cast<uint32_t>(handle) >> kIndexBits);
-            if (index >= N or not used_[index] or gen_[index] != gen)
+            // Full high bits, not truncated to 16: a handle carrying junk above the generation
+            // field (bits set beyond what handle_for ever produces) must fail to resolve, not alias.
+            uint32_t const gen = static_cast<uint32_t>(handle) >> kIndexBits;
+            if (index >= N or not used_[index] or static_cast<uint32_t>(gen_[index]) != gen)
             {
                 return nullptr;
             }
